// lib/api/types.gen.ts
export type IncidentEditModel = {
    startDate: string;
    stopDate: string;
    aCrew: number;
    bCrew: number;
    cCrew: number;
    dCrew: number;
    dayCrew: number;
};
// ...
export type ProductionBonusEditModel = {
    id?: number;
    startDate?: string;
    stopDate?: string;
    aCrew?: number;
    bCrew?: number;
    cCrew?: number;
    dCrew?: number;
    average?: number;
};

export type RunningBonusModel = {
    crew: string;
    runningBonus: number;
    totalHours: number;
};